Re-added legacy support for binary serialization in the move to .NET Standard.
Previously this step was throwing a SerializationException, because the CodeNamespace class (now in a Nuget package) used in creating .iip files is no longer marked with the [Serializable] attribute, which is required for use with the BinaryFormatter. It should also be noted that this exception is only thrown with a clean output directory (and not subsequent builds), as the code doesn't treat the exception as crashing but generates the needed Kernel.cs file anyway (which skips all this on subsequent builds). Regardless, the solution, in lieu of removing this component altogether, was to add a custom SurrogateSelector (in added package KGySoft.CoreLibraries) to the BinaryFormatter which allows it to serialize even objects not marked with [Serializable]. It should be noted that the output files (.iip and .json) are similar to the original 4.6.1 files but not identical. It should also be carefully noted that .NET is moving, and advising moving, away from the use of BinaryFormatter for security reasons (https://docs.microsoft.com/en-us/dotnet/standard/serialization/binaryformatter-security-guide). With this in mind, it's highly possible that a better solution would in fact be to remove this component and support only the safer JSON serialization already present.