Small improvements to nuget package
These changes allow me to build the nuget packages locally. This change
also modifies the output directory for `IdlImpTool` - using the same
directory causes errors when creating the package because
`IDLImporter.xml` is found in both the original location in the source
tree as well as in the output directory.